APF-1356 Adressed review comments.
