+++ connectors/concur/src/main/java/com/vmware/connectors/concur/ConcurController.java
@@ -19,7 +19,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
-import org.springframework.http.CacheControl;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.LinkedMultiValueMap;
@@ -35,13 +34,11 @@ import reactor.core.publisher.Mono;
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.io.IOException;
-import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
 import java.util.Locale;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
 
 import static com.vmware.connectors.common.utils.CommonUtils.APPROVAL_ACTIONS;
 import static com.vmware.connectors.concur.ConcurConstants.ConcurRequestActions.*;
@@ -58,9 +55,6 @@ public class ConcurController {
     private static final Logger logger = LoggerFactory.getLogger(ConcurController.class);
 
     private final WebClient rest;
-    private final String metadata;
-    private final long maxAge;
-    private final TimeUnit unit;
     private final CardTextAccessor cardTextAccessor;
     private final Resource concurrRequestTemplate;
 
@@ -83,26 +77,13 @@ public class ConcurController {
                             @Value("${concur.client-id}") final String clientId,
                             @Value("${concur.client-secret}") final String clientSecret,
                             @Value("${concur.oauth-instance-url}") final String oauthTokenUrl,
-                            @Value("classpath:static/templates/concur-request-template.xml") Resource concurRequestTemplate,
-                            @Value("classpath:static/discovery/metadata.json") Resource metadataJsonResource,
-                            @Value("${rootDiscovery.cacheControl.maxAge:1}") long maxAge,
-                            @Value("${rootDiscovery.cacheControl.unit:HOURS}") TimeUnit unit) throws IOException {
+                            @Value("classpath:static/templates/concur-request-template.xml") Resource concurRequestTemplate) {
         this.rest = rest;
         this.cardTextAccessor = cardTextAccessor;
         this.clientId = clientId;
         this.clientSecret = clientSecret;
         this.oauthTokenUrl = oauthTokenUrl;
         this.concurrRequestTemplate = concurRequestTemplate;
-        this.metadata = IOUtils.toString(metadataJsonResource.getInputStream(), Charset.defaultCharset());
-        this.maxAge = maxAge;
-        this.unit = unit;
-    }
-
-    @GetMapping(path = "/")
-    public ResponseEntity<String> getMetadata(HttpServletRequest request) {
-        return ResponseEntity.ok()
-                .cacheControl(CacheControl.maxAge(maxAge, unit))
-                .body(this.metadata.replace("${CONNECTOR_HOST}", CommonUtils.buildConnectorUrl(request, null)));
     }
 
     @PostMapping(path = "/cards/requests",
