+++ connectors/jira/src/main/java/com/vmware/connectors/jira/JiraController.java
@@ -12,14 +12,10 @@ import com.vmware.connectors.common.payloads.response.*;
 import com.vmware.connectors.common.utils.CardTextAccessor;
 import com.vmware.connectors.common.utils.CommonUtils;
 import com.vmware.connectors.common.utils.Reactive;
-import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.Resource;
-import org.springframework.http.CacheControl;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -31,10 +27,7 @@ import reactor.core.publisher.Mono;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
-import java.io.IOException;
-import java.nio.charset.Charset;
 import java.util.*;
-import java.util.concurrent.TimeUnit;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpMethod.GET;
@@ -53,29 +46,13 @@ public class JiraController {
 
     private static final int COMMENTS_SIZE = 2;
 
-    private final String metadata;
     private final WebClient rest;
     private final CardTextAccessor cardTextAccessor;
-    private final long maxAge;
-    private final TimeUnit unit;
 
     @Autowired
-    public JiraController(WebClient rest, CardTextAccessor cardTextAccessor,
-                          @Value("classpath:static/discovery/metadata.json") Resource metadataJsonResource,
-                          @Value("${rootDiscovery.cacheControl.maxAge:1}") long maxAge,
-                          @Value("${rootDiscovery.cacheControl.unit:HOURS}") TimeUnit unit) throws IOException {
+    public JiraController(WebClient rest, CardTextAccessor cardTextAccessor) {
         this.rest = rest;
         this.cardTextAccessor = cardTextAccessor;
-        this.metadata = IOUtils.toString(metadataJsonResource.getInputStream(), Charset.defaultCharset());
-        this.maxAge = maxAge;
-        this.unit = unit;
-    }
-
-    @GetMapping(path = "/")
-    public ResponseEntity<String> getMetadata(HttpServletRequest request) {
-        return ResponseEntity.ok()
-                .cacheControl(CacheControl.maxAge(maxAge, unit))
-                .body(this.metadata.replace("${CONNECTOR_HOST}", CommonUtils.buildConnectorUrl(request, null)));
     }
 
     @PostMapping(path = "/cards/requests", produces = APPLICATION_JSON_VALUE, consumes = APPLICATION_JSON_VALUE)
